// src/utils/api/cms/cms.enum.ts
export enum PlacementDisplayType {
  DOUBLE_BANNER = "DOUBLE_BANNER",
  CAROUSEL = "CAROUSEL",
  SLIDER = "SLIDER",
  POPUP = "POPUP",
  GRID = "GRID",
}

export enum BlockTypeCode {
  BANNER = "BANNER",
  PRODUCT_CAROUSEL = "PRODUCT_CAROUSEL",
  INFO_CARDS = "INFO_CARDS",
  IMAGE_GALLERY = "IMAGE_GALLERY",
  NEWSLETTER_SIGNUP = "NEWSLETTER_SIGNUP",
  PRODUCT_CATEGORY_NAV = "PRODUCT_CATEGORY_NAV",
  PRODUCT_LIST = "PRODUCT_LIST",
  PRODUCT_COLLECTION_SHOWCASE = "PRODUCT_COLLECTION_SHOWCASE",
  PRODUCT_EXPANDABLE_DESCRIPTION = "PRODUCT_EXPANDABLE_DESCRIPTION",
  TEXT_BLOCK = "TEXT_BLOCK",
  JSON_DISPLAY = "JSON_DISPLAY",
}

export enum ActionType {
  NONE = "NONE",
  LINK = "LINK",
}

export enum AnimationType {
  STATIC = "STATIC",
  MARQUEE = "MARQUEE",
}