// src/utils/api/banner/banner.enum.ts
export enum MediaType {
  IMAGE = "IMAGE",
  VIDEO = "VIDEO",
  GIF = "GIF",
}

export enum MobileOddFullWidthItem {
  FIRST = "FIRST",
  LAST = "LAST",
}

export enum BannerLayout {
  RIGHT_TOP = "RIGHT_TOP",
  CENTER_TOP = "CENTER_TOP",
  LEFT_TOP = "LEFT_TOP",

  RIGHT_CENTER = "RIGHT_CENTER",
  CENTER = "CENTER",
  LEFT_CENTER = "LEFT_CENTER",

  RIGHT_BOTTOM = "RIGHT_BOTTOM",
  CENTER_BOTTOM = "CENTER_BOTTOM",
  LEFT_BOTTOM = "LEFT_BOTTOM",
}